tests/sample_compat: Make list_samples() public (within the crate)
That will be helpful to write other tests using the samples.
