+++ tests/bin/bsdiff
